[java] Move rule SignatureDeclareThrowsException from typeresolution to strictexception Replace existing rule with the typeresolution-based implementation.
