+++ pmd-core/src/main/java/net/sourceforge/pmd/RuleSetFactoryCompatibility.java
@@ -74,6 +74,7 @@ public class RuleSetFactoryCompatibility {
         addFilterRuleMoved("java", "typeresolution", "coupling", "LooseCoupling");
         addFilterRuleMoved("java", "typeresolution", "clone", "CloneMethodMustImplementCloneable");
         addFilterRuleMoved("java", "typeresolution", "imports", "UnusedImports");
+        addFilterRuleMoved("java", "typeresolution", "strictexception", "SignatureDeclareThrowsException");
     }
 
     void addFilterRuleRenamed(String language, String ruleset, String oldName, String newName) {
