+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/strictexception/SignatureDeclareThrowsExceptionRule.java
@@ -9,15 +9,27 @@ import java.util.List;
 
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.java.ast.ASTAnnotation;
+import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceType;
 import net.sourceforge.pmd.lang.java.ast.ASTCompilationUnit;
 import net.sourceforge.pmd.lang.java.ast.ASTConstructorDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTExtendsList;
+import net.sourceforge.pmd.lang.java.ast.ASTImplementsList;
 import net.sourceforge.pmd.lang.java.ast.ASTImportDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTName;
 import net.sourceforge.pmd.lang.java.ast.ASTNameList;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRule;
+import net.sourceforge.pmd.properties.BooleanProperty;
 
 /**
+ * A method/constructor shouldn't explicitly throw java.lang.Exception, since it
+ * is unclear which exceptions that can be thrown from the methods. It might be
+ * difficult to document and understand such vague interfaces. Use either a class
+ * derived from RuntimeException or a checked exception.
+ * 
+ * <p>This rule uses PMD's type resolution facilities, and can detect
+ * if the class implements or extends TestCase class
  *
  * @author <a href="mailto:trondandersen@c2i.net">Trond Andersen</a>
  * @version 1.0
@@ -26,12 +38,78 @@ import net.sourceforge.pmd.lang.java.rule.AbstractJavaRule;
 
 public class SignatureDeclareThrowsExceptionRule extends AbstractJavaRule {
 
-    private boolean junitImported;
+    private static final BooleanProperty IGNORE_JUNIT_COMPLETELY_DESCRIPTOR = new BooleanProperty(
+            "IgnoreJUnitCompletely", "Allow all methods in a JUnit testcase to throw Exceptions", false, 1.0f);
+
+    // Set to true when the class is determined to be a JUnit testcase
+    private boolean junitImported = false;
+
+    public SignatureDeclareThrowsExceptionRule() {
+        definePropertyDescriptor(IGNORE_JUNIT_COMPLETELY_DESCRIPTOR);
+    }
 
     @Override
     public Object visit(ASTCompilationUnit node, Object o) {
         junitImported = false;
         return super.visit(node, o);
+    }
+
+    @Override
+    public Object visit(ASTClassOrInterfaceDeclaration node, Object data) {
+        if (junitImported) {
+            return super.visit(node, data);
+        }
+
+        ASTImplementsList impl = node.getFirstChildOfType(ASTImplementsList.class);
+        if (impl != null && impl.jjtGetParent().equals(node)) {
+            for (int ix = 0; ix < impl.jjtGetNumChildren(); ix++) {
+                Node child = impl.jjtGetChild(ix);
+
+                if (child.getClass() != ASTClassOrInterfaceType.class) {
+                    continue;
+                }
+
+                ASTClassOrInterfaceType type = (ASTClassOrInterfaceType) child;
+                if (isJUnitTest(type)) {
+                    junitImported = true;
+                    return super.visit(node, data);
+                }
+            }
+        }
+        if (node.jjtGetNumChildren() != 0 && node.jjtGetChild(0) instanceof ASTExtendsList) {
+            ASTClassOrInterfaceType type = (ASTClassOrInterfaceType) node.jjtGetChild(0).jjtGetChild(0);
+            if (isJUnitTest(type)) {
+                junitImported = true;
+                return super.visit(node, data);
+            }
+        }
+
+        return super.visit(node, data);
+    }
+
+    private boolean isJUnitTest(ASTClassOrInterfaceType type) {
+        Class<?> clazz = type.getType();
+        if (clazz == null) {
+            if ("junit.framework.Test".equals(type.getImage())) {
+                return true;
+            }
+        } else if (isJUnitTest(clazz)) {
+            return true;
+        } else {
+            while (clazz != null && !Object.class.equals(clazz)) {
+                for (Class<?> intf : clazz.getInterfaces()) {
+                    if (isJUnitTest(intf)) {
+                        return true;
+                    }
+                }
+                clazz = clazz.getSuperclass();
+            }
+        }
+        return false;
+    }
+
+    private boolean isJUnitTest(Class<?> clazz) {
+        return clazz.getName().equals("junit.framework.Test");
     }
 
     @Override
@@ -44,8 +122,7 @@ public class SignatureDeclareThrowsExceptionRule extends AbstractJavaRule {
 
     @Override
     public Object visit(ASTMethodDeclaration methodDeclaration, Object o) {
-        if ((methodDeclaration.getMethodName().equals("setUp") || methodDeclaration.getMethodName().equals("tearDown"))
-                && junitImported) {
+        if (junitImported && isAllowedMethod(methodDeclaration)) {
             return super.visit(methodDeclaration, o);
         }
 
@@ -62,24 +139,39 @@ public class SignatureDeclareThrowsExceptionRule extends AbstractJavaRule {
             }
         }
 
+        checkExceptions(methodDeclaration, o);
+
+        return super.visit(methodDeclaration, o);
+    }
+
+    private boolean isAllowedMethod(ASTMethodDeclaration methodDeclaration) {
+        if (getProperty(IGNORE_JUNIT_COMPLETELY_DESCRIPTOR)) {
+            return true;
+        } else {
+            return methodDeclaration.getMethodName().equals("setUp")
+                    || methodDeclaration.getMethodName().equals("tearDown");
+        }
+    }
+
+    @Override
+    public Object visit(ASTConstructorDeclaration constructorDeclaration, Object o) {
+        checkExceptions(constructorDeclaration, o);
+
+        return super.visit(constructorDeclaration, o);
+    }
+
+    /**
+     * Search the list of thrown exceptions for Exception
+     */
+    private void checkExceptions(Node method, Object o) {
         List<ASTName> exceptionList = Collections.emptyList();
-        ASTNameList nameList = methodDeclaration.getFirstChildOfType(ASTNameList.class);
+        ASTNameList nameList = method.getFirstChildOfType(ASTNameList.class);
         if (nameList != null) {
             exceptionList = nameList.findDescendantsOfType(ASTName.class);
         }
         if (!exceptionList.isEmpty()) {
             evaluateExceptions(exceptionList, o);
         }
-        return super.visit(methodDeclaration, o);
-    }
-
-    @Override
-    public Object visit(ASTConstructorDeclaration constructorDeclaration, Object o) {
-        List<ASTName> exceptionList = constructorDeclaration.findDescendantsOfType(ASTName.class);
-        if (!exceptionList.isEmpty()) {
-            evaluateExceptions(exceptionList, o);
-        }
-        return super.visit(constructorDeclaration, o);
     }
 
     /**
