+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/strictexception/MyInterface.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.rule.typeresolution.xml;
+package net.sourceforge.pmd.lang.java.rule.strictexception;
 
 /**
  * Warning, this class ARE not useless. It is used by the some of regression
