--- pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/typeresolution/SignatureDeclareThrowsExceptionTest.java
@@ -1,17 +1,0 @@
-/**
- * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
- */
-
-package net.sourceforge.pmd.lang.java.rule.typeresolution;
-
-import net.sourceforge.pmd.testframework.SimpleAggregatorTst;
-
-public class SignatureDeclareThrowsExceptionTest extends SimpleAggregatorTst {
-
-    private static final String RULESET = "java-typeresolution";
-
-    @Override
-    public void setUp() {
-        addRule(RULESET, "SignatureDeclareThrowsException");
-    }
-}
